+++ src/index.js
@@ -1,6 +1,22 @@
-// En este archivo definirás tus rutas e importarás los componentes que vas a renderizar.
-import { testingRouter } from './router.js';
+import { Home } from './views/Home.js';
+import { chatGrupal } from './views/chatGrupal.js';
+// ... import other views
+import { setRootEl, setRoutes, onURLChange } from './router.js';
 
-document.addEventListener('DOMContentLoaded', () => {
-  testingRouter();
+// Define your routes and their associated views
+const routes = {
+  '/': Home,
+  '/chatGrupal': chatGrupal
+};
+setRoutes(routes); // Assign the routes
+
+
+// Set the root element where views will be rendered
+
+window.addEventListener('DOMContentLoaded', () => { //POR QUE EN LA GUIA ESTA WINDOW 
+  const rootElement = document.getElementById("root");
+  //const header ???
+  //constfooter ??
+  setRootEl(rootElement);
+  onURLChange();
 });
