+++ src/router.js
@@ -1,27 +1,72 @@
-import { Home } from './views/Home.js';
-import { chatGrupal } from './views/groupChat.js';
+//import { Home } from './views/Home.js';
+//import { chatGrupal } from './views/groupChat.js';  // ... import other views
 
-const routes = {
-  '/': Home,
-  '/chatGrupal': chatGrupal
+//const routes = {
+//'/': Home,
+//'/chatGrupal': chatGrupal
+//};
+
+//const rootDiv = document.getElementById('root');
+
+//export function testingRouter() {
+//window.addEventListener('hashchange', () => {
+//loadRoute();
+//});
+
+//loadRoute(); // Cargar la ruta inicial
+//}
+
+//function loadRoute() {
+//const path = window.location.hash.replace('#', '') || '/';
+//const view = routes[path];
+//if (view) {
+// rootDiv.innerHTML = '';
+//rootDiv.appendChild(view());
+//}
+//}
+
+let ROUTES = {};
+let rootEl = "";
+
+export const setRootEl = (el) => {
+  rootEl = el; //assing rootel
+};
+export const setRoutes = (routes) => {
+  ROUTES = routes; //assing routes
 };
 
-const rootDiv = document.getElementById('root');
+const queryStringToObject = (queryString) => {
+  const params = new URLSearchParams(queryString);   // convert query string to URLSearchParams
 
-export function testingRouter() {
-  window.addEventListener('hashchange', () => {
-    loadRoute();
-  });
+  return Object.fromEntries(params.entries());   // convert URLSearchParams to an object and return the object
 
-  loadRoute(); // Cargar la ruta inicial
-}
+};
 
-function loadRoute() {
-  const path = window.location.hash.replace('#', '') || '/';
-  const view = routes[path];
+const renderView = (pathName, props = {}) => {
+  rootEl.innerHTML = "";   // clear the root element
 
-  if (view) {
-    rootDiv.innerHTML = '';
-    rootDiv.appendChild(view());
+
+  const viewFunc = ROUTES[pathName];   // find the correct view in ROUTES for the pathname
+
+  if (!viewFunc) {
+    return navigateTo("/page/404");   // in case not found render the error view
+
   }
-}
+  const viewEl = viewFunc(props);   // render the correct view passing the value of props
+
+  rootEl.appendChild(viewEl);  // add the view element to the DOM root element
+
+};
+
+export const navigateTo = (pathname, props = {}) => {
+  window.history.pushState({}, pathname, window.location.origin + pathname);  // update window history with pushState
+
+  renderView(pathname, props);  // render the view with the pathname and props
+
+};
+
+export const onURLChange = () => {
+  const { pathname, search } = window.location;// parse the location for the pathname and search params
+  const props = queryStringToObject(search); // convert the search params to an object
+  renderView(pathname, props); // render the view with the pathname and object
+};
